move util to internal
