+++ launcher/spec/launch_spec.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"cloud.google.com/go/compute/metadata"
+	"github.com/google/go-tpm-tools/internal/util"
 	"github.com/google/go-tpm-tools/launcher/internal/experiments"
-	"github.com/google/go-tpm-tools/verifier/util"
 )
 
 // RestartPolicy is the enum for the container restart policy.
